Conflicts:
	FallRiver/FallRiver/BaseCharacter.cpp
	FallRiver/FallRiver/BaseObject.cpp
	FallRiver/FallRiver/Boss1.cpp
	FallRiver/FallRiver/Boss2.cpp
	FallRiver/FallRiver/Bullet.cpp
	FallRiver/FallRiver/ChasingAI.cpp
	FallRiver/FallRiver/CompanionAI.cpp
	FallRiver/FallRiver/DirectInput.cpp
	FallRiver/FallRiver/Enemy.cpp
	FallRiver/FallRiver/HUD.cpp
	FallRiver/FallRiver/HighScoresMenuState.cpp
	FallRiver/FallRiver/Level.cpp
	FallRiver/FallRiver/LoadMenuState.cpp
	FallRiver/FallRiver/LoseMenuState.cpp
	FallRiver/FallRiver/NPC.cpp
	FallRiver/FallRiver/ObjectManager.cpp
	FallRiver/FallRiver/PauseMenuState.cpp
	FallRiver/FallRiver/Player.cpp
	FallRiver/FallRiver/ShootingAi.cpp
	FallRiver/FallRiver/SpawnPoint.cpp
	FallRiver/FallRiver/Terrain.cpp
	FallRiver/FallRiver/TutorialState.cpp
	FallRiver/FallRiver/ViewManager.cpp
	FallRiver/FallRiver/Weapon.cpp
	FallRiver/FallRiver/WinMenuState.cpp

+++ FallRiver/FallRiver/Boss1.cpp
@@ -380,14 +380,14 @@ void Boss1::Update(float fElapsedTime)
 	{
 		m_playerAnim.fTime -= thisAnim->frames[m_playerAnim.curAnimation][m_playerAnim.curFrame].duration;
 		m_playerAnim.curFrame++;
-		if(m_playerAnim.curFrame < (int)thisAnim->frames[m_playerAnim.curAnimation].size())
+		if((unsigned int)m_playerAnim.curFrame < (int)thisAnim->frames[m_playerAnim.curAnimation].size())
 		{
 			if(strcmp(thisAnim->frames[m_playerAnim.curAnimation][m_playerAnim.curFrame].eventMsg,"none") != 0)
 				EventSystem::GetInstance()->SendEvent(thisAnim->frames[m_playerAnim.curAnimation][m_playerAnim.curFrame].eventMsg, this);
 		}
-		if(((unsigned)m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size()) && thisAnim->looping[m_playerAnim.curAnimation])
+		if(((unsigned int)m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size()) && thisAnim->looping[m_playerAnim.curAnimation])
 			m_playerAnim.curFrame = 0;
-		else if((unsigned)m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size() && !thisAnim->looping[m_playerAnim.curAnimation])
+		else if((unsigned int)m_playerAnim.curFrame == thisAnim->frames[m_playerAnim.curAnimation].size() && !thisAnim->looping[m_playerAnim.curAnimation])
 			m_playerAnim.curFrame--;
 	}
 }
